+++ alacritty_terminal/src/ansi.rs
@@ -12,6 +12,7 @@ use vte::{Params, ParamsIter};
 use alacritty_config_derive::ConfigDeserialize;
 
 use crate::index::{Column, Line};
+use crate::term::cell::Hyperlink;
 use crate::term::color::Rgb;
 
 /// Maximum time before a synchronized update is aborted.
@@ -457,6 +458,9 @@ pub trait Handler {
 
     /// Report text area size in characters.
     fn text_area_size_chars(&mut self) {}
+
+    /// Set hyperlink.
+    fn set_hyperlink(&mut self, _: Option<Hyperlink>) {}
 }
 
 /// Terminal cursor configuration.
@@ -1005,6 +1009,27 @@ where
                         unhandled(params);
                     }
                 }
+            },
+
+            // Hyperlink.
+            b"8" if params.len() > 2 => {
+                let link_params = params[1];
+                let uri = str::from_utf8(params[2]).unwrap_or_default();
+
+                // The OSC 8 escape sequence must be stopped when getting an empty `uri`.
+                if uri.is_empty() {
+                    self.handler.set_hyperlink(None);
+                    return;
+                }
+
+                // Link parameters are in format of `key1=value1:key2=value2`. Currently only key
+                // `id` is defined.
+                let id = link_params
+                    .split(|&b| b == b':')
+                    .find_map(|kv| kv.strip_prefix(b"id="))
+                    .and_then(|kv| str::from_utf8(kv).ok());
+
+                self.handler.set_hyperlink(Some(Hyperlink::new(id, uri)));
             },
 
             // Get/set Foreground, Background, Cursor colors.